Move NanopubClient to its own module client.py
