+++ nanopub/__init__.py
@@ -1,2 +1,3 @@
-from .nanopub import NanopubClient, Nanopub
+from .nanopub import Nanopub
+from .client import NanopubClient
 from ._version import __version__
